requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
autopep8==2.3.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.0
defusedxml==0.8.0rc2
Django==5.0.7
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-filter==24.2
django-mail-templated==2.6.5
django-redis==5.4.0
django-silk==5.1.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
drf-nested-routers==0.94.1
environs==11.0.0
flower==2.0.1
gprof2dot==2024.6.6
gunicorn==22.0.0
humanize==4.10.0
idna==3.7
kombu==5.3.7
marshmallow==3.21.3
mysqlclient==2.2.4
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pycodestyle==2.12.0
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
redis==5.0.7
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomli==2.0.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
